third_party/blink/web_tests/virtual/deprecate-unload/external/wpt/html/browsers/browsing-the-web/unloading-documents/004-expected.txt

This is a testharness.js-based test.
[FAIL] document.open in beforeunload with button
  assert_equals: expected "ABCDEFGHIJKLMN" but got "ABGLMN"
Harness: the test ran to completion.
